+++ gossipd/test/run-bench-find_route.c
@@ -131,9 +131,10 @@ static void add_connection(struct routing_state *rstate,
 	memcpy((char *)&scid + (!idx) * sizeof(to), &to, sizeof(to));
 
 	chan = get_channel(rstate, &scid);
-	if (!chan)
+	if (!chan) {
 		chan = new_chan(rstate, &scid, &nodes[from], &nodes[to],
 				AMOUNT_SAT(1000000));
+	}
 
 	c = &chan->half[idx];
 	c->base_fee = base_fee;
@@ -217,7 +218,7 @@ int main(int argc, char *argv[])
 	setup_tmpctx();
 
 	me = nodeid(0);
-	rstate = new_routing_state(tmpctx, NULL, &me, 0, NULL, NULL, NULL);
+	rstate = new_routing_state(tmpctx, NULL, &me, 0, NULL, NULL);
 	opt_register_noarg("--perfme", opt_set_bool, &perfme,
 			   "Run perfme-start and perfme-stop around benchmark");